% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{cleanGPS_trip}
\alias{cleanGPS_trip}
\title{\code{cleanGPS} cleans raw GPS data}
\usage{
cleanGPS_trip(trip, trip_id = NULL, min.gps.obs = 50,
  max.idle.time = 4, max.time.between.gps = 120, min.speed = 5,
  median.speed = 20, minmax.speed = 30, min.distance = 1000,
  min.endpoint.speed = 10, tag = NULL, verbose = FALSE)
}
\description{
\code{cleanGPS} cleans raw GPS data
}
\details{
Cleans raw GPS data by removing trips with outside the indicated parameters.
}
