% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_param_by_factor.R
\name{gaussian_param_by_factor}
\alias{gaussian_param_by_factor}
\title{Calculate the parameters of Gaussian speeds}
\usage{
gaussian_param_by_factor(logspeeds, byFactor, states = NULL)
}
\arguments{
\item{logspeeds}{A numeric vector of speed observations (in km/h) on the (natural) log-scale, of size \code{n = number of observations}.}

\item{byFactor}{A vector of factors (link + time bin) of size \code{n}, i.e. one factor for each speed observation.}

\item{state_prob}{An \code{n x nQ} matrix of smoothed state probabilities where \code{nQ = number of states}.}
}
\value{
\code{gaussian_param_by_factor} returns a list of two matrices of identical dimensions \code{m X nQ}
where \code{m = number of levels} (links + timeBins).  The first list \code{mean} contains the vector of means
for each level and state.  The second list \code{sigma2} contains the vector of variances for each level and state.
}
\description{
\code{gaussian_param_by_factor} calculates the mean and variance of Gaussian speeds.  The function refers to the calculation
of \eqn{\mu} and \eqn{\sigma^{2}} at step 4 of Algorithm 1 in Woodard et al., 2017
}
\references{
{Woodard, D., Nogin, G., Koch, P., Racz, D., Goldszmidt, M., Horvitz, E., 2017.  Predicting travel time reliability using mobile phone GPS data.  Transportation Research Part C, 75, 30-44.}
}
