% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getValidE.R
\name{getValidE}
\alias{getValidE}
\title{Get an appropriate value for \code{logE}}
\usage{
getValidE(object, logE, n)
}
\arguments{
\item{object}{A model object (a list) provided through the execution of function \code{timetravelHMM}.}

\item{logE}{Tentative values of point estimates of trip effects.}

\item{n}{Number of samples. Default is 1000.}
}
\value{
The function return a valid value for \code{logE}.  The value supplied as a parameter
  for \code{logE} is a numerical vector of size \code{n}, in which case it is
  returned \emph{as is}.  If a single numeric value is supplied, it will be replicated into
  a vector of the appropriate size which will be returned.  If \code{logE} is \code{NULL}
  or invalid, then the function will return either a vector of simulated values
  (if the model is from the \code{trip} family), or a vector of \code{0} otherwise.
}
\description{
Get an appropriate value for \code{logE}
}
\keyword{Function}
\keyword{The}
\keyword{\code{getValidE}}
\keyword{\code{logE}}
\keyword{\emph{NOT}}
\keyword{a}
\keyword{and}
\keyword{as}
\keyword{for}
\keyword{in}
\keyword{internal}
\keyword{is}
\keyword{it.}
\keyword{model}
\keyword{object}
\keyword{parameter}
\keyword{passed}
\keyword{returns}
\keyword{takes}
\keyword{tentative}
\keyword{the}
\keyword{used.}
\keyword{valid}
\keyword{value}
