% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{normalizeR}
\alias{normalizeR}
\title{Normalize each row of a matrix by the sum of its elements}
\usage{
normalizeR(x, m, n)
}
\arguments{
\item{x}{A matrix of dimensions \code{m X n}.}

\item{m}{The number of rows of the matrix.}

\item{n}{The number of columns of the matrix.}
}
\value{
A matrix of dimensions \code{m X n} with each row summing up to 1.
}
\description{
Normalize each row of a matrix by the sum of its elements
}
\examples{
normalizeR(matrix(c(1,2,3,4), nrow=2, ncol = 2), 2, 2)
}
\keyword{Function}
\keyword{\code{normalizeR}}
\keyword{a}
\keyword{by}
\keyword{each}
\keyword{elements.}
\keyword{internal}
\keyword{its}
\keyword{matrix}
\keyword{normalizes}
\keyword{of}
\keyword{row}
\keyword{sum}
\keyword{the}
