% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{normalizeV}
\alias{normalizeV}
\title{Normalize a vector by the sum of its elements}
\usage{
normalizeV(x)
}
\arguments{
\item{x}{A vector of dimension \code{n}.}
}
\value{
A vector of dimension \code{n} with  its elements summing up to 1.
}
\description{
Normalize a vector by the sum of its elements
}
\examples{
\dontrun{
normalizeV(c(1,3))
normalizeV(matrix(c(1,2,3,4), nrow=2, ncol = 2), 2, 2)
}
}
\keyword{.}
\keyword{.colSums(x,}
\keyword{Function}
\keyword{It}
\keyword{\code{normalizeV}}
\keyword{a}
\keyword{by}
\keyword{elements.}
\keyword{faster}
\keyword{internal}
\keyword{is}
\keyword{its}
\keyword{m=1,}
\keyword{much}
\keyword{n=length(x))}
\keyword{normalizes}
\keyword{of}
\keyword{or}
\keyword{sum}
\keyword{than}
\keyword{the}
\keyword{vector}
