% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{order_states}
\alias{order_states}
\title{Order manually the states vector}
\usage{
order_states(param)
}
\arguments{
\item{param}{...}
}
\value{
A list of indeces to sort by and how to sort them
}
\description{
Order manually the states vector
}
\keyword{Function}
\keyword{\code{order_states}}
\keyword{a}
\keyword{based}
\keyword{each}
\keyword{internal}
\keyword{mean.}
\keyword{of}
\keyword{on}
\keyword{orders}
\keyword{row}
\keyword{states}
\keyword{the}
\keyword{vector}
