% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.traveltime.no_dependence}
\alias{predict.traveltime.no_dependence}
\title{Predict the travel time for a trip using a \code{traveltimeHMM} model object that is not of the HMM family}
\usage{
\method{predict}{traveltime.no_dependence}(object, tripdata, starttime,
  n = 1000, logE = NULL, ...)
}
\arguments{
\item{object}{A model object (a list) provided through the execution of
function \code{timetravelHMM} for a \code{trip} or \code{no-dependence} model type.
The list includes information on model as well as estimates
for its parameters.  See \code{timetravelHMM} man page.}

\item{tripdata}{A data frame of road links with information on each
link's traversal.  Columns minimally includes objects 'linkID' and 'length',
and the latter must have the same length.  Rows must be in chronological order.}

\item{starttime}{The start date and time for the very first link of the trip, in POSIXct format.}

\item{n}{Number of samples. Default is 1000.}

\item{logE}{Point estimate of trip effects, in the form of a numerical vector of size \code{n}.}

\item{...}{not used.}
}
\value{
\code{predict.traveltime.no_dependence} returns a vector of size \code{n} of representing the point prediction of total travel time, in seconds, for each run.
}
\description{
Predict the travel time for a trip using a \code{traveltimeHMM} model object that is not of the HMM family
}
\references{
{Woodard, D., Nogin, G., Koch, P., Racz, D., Goldszmidt, M., Horvitz, E., 2017.  Predicting travel time reliability using mobile phone GPS data.  Transportation Research Part C, 75, 30-44.}
}
\keyword{(see}
\keyword{2}
\keyword{2017.}
\keyword{Algorithm}
\keyword{However,}
\keyword{The}
\keyword{Woodard}
\keyword{\code{HMM}}
\keyword{\code{predict.traveltime.no_dependence}}
\keyword{\code{predict.traveltime}).}
\keyword{\code{traveltimeHMM}}
\keyword{a}
\keyword{al.,}
\keyword{and}
\keyword{are}
\keyword{as}
\keyword{at}
\keyword{by}
\keyword{estimates}
\keyword{estimation}
\keyword{et}
\keyword{family}
\keyword{for}
\keyword{from}
\keyword{function}
\keyword{generated}
\keyword{handled}
\keyword{implements}
\keyword{initial}
\keyword{internal}
\keyword{is}
\keyword{man}
\keyword{matrix}
\keyword{model}
\keyword{not}
\keyword{object}
\keyword{of}
\keyword{page}
\keyword{parameter}
\keyword{performs}
\keyword{point}
\keyword{prediction}
\keyword{probability}
\keyword{provided}
\keyword{simulation}
\keyword{stage.}
\keyword{state}
\keyword{that}
\keyword{the}
\keyword{they}
\keyword{transition}
\keyword{using}
\keyword{vector}
\keyword{were}
