% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_bins.R
\name{time_bins}
\alias{time_bins}
\title{A mapping from real time to time bins}
\usage{
time_bins(t)
}
\arguments{
\item{t}{A real time `POSIXlt` time stamp}
}
\value{
A string as a time bin
}
\description{
\code{time_bins} transforms a real time to a time bin
}
\details{
...
}
