% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_bins.R
\name{time_bins_functional}
\alias{time_bins_functional}
\title{Transforms a list of rules to a functional}
\usage{
time_bins_functional(
  time_bin_readable_function = time_bins_readable,
  period = c("hours", "minutes")
)
}
\arguments{
\item{rules}{...}
}
\value{
...
}
\description{
Transforms a list of rules to a functional
}
\details{
...
}
\keyword{\code{time_bins_functional}}
\keyword{a}
\keyword{bin}
\keyword{for}
\keyword{function}
\keyword{functional}
\keyword{human}
\keyword{internal}
\keyword{mapping}
\keyword{of}
\keyword{performance}
\keyword{readable}
\keyword{rules}
\keyword{time}
\keyword{to}
\keyword{transforms}
