% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_bins.R
\name{time_bins_readable}
\alias{time_bins_readable}
\title{A simple example of time bins function that is human readable.}
\usage{
time_bins_readable(t)
}
\arguments{
\item{t}{A datetime stamp.}
}
\value{
a string representing time bin out of \code{MorningRush}, \code{EveningRush}, \code{EveningNight}, \code{Weekendday}, \code{Weekday}.
}
\description{
A simple example of time bins function that is human readable.
}
\details{
...
}
\examples{
time_bins_readable(Sys.time())
}
\keyword{\code{time_bins_readable}}
\keyword{a}
\keyword{bin}
\keyword{converts}
\keyword{datetime}
\keyword{internal}
\keyword{stamp}
\keyword{tag}
\keyword{time}
\keyword{to}
