% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm_estimates.R
\name{tmat_est}
\alias{tmat_est}
\title{Compute transition matrix probabilities}
\usage{
tmat_est(joint_prob, state_prob, init_ids, by_factor)
}
\arguments{
\item{joint_prob}{An \code{n-1 x Q^2} matrix of joint state probabilities for \code{n} observations and \code{Q} states}

\item{state_prob}{An \code{n-1 x Q} matrix of initial state probabilities for \code{n} observations and \code{Q} states}

\item{init_ids}{A vector of indices of the initial state observation for each trip}

\item{by_factor}{A vector of factors (link + time bin) of size \code{n} (a factor for each row of \code{state_prob}).}
}
\value{
An \code{m x Q} matrix of probabilites for the m levels of \code{by_factor}.
}
\description{
Compute transition matrix probabilities
}
\keyword{1}
\keyword{2017}
\keyword{4}
\keyword{Algorithm}
\keyword{Big_Gamma_j,b^(t+1)(q',}
\keyword{These}
\keyword{Woodard}
\keyword{\code{by_factor}.}
\keyword{\code{tmat_est}}
\keyword{a.,}
\keyword{at}
\keyword{correspond}
\keyword{each}
\keyword{et}
\keyword{for}
\keyword{in}
\keyword{internal}
\keyword{level}
\keyword{of}
\keyword{passed}
\keyword{probabilities}
\keyword{q)}
\keyword{returns}
\keyword{state}
\keyword{step}
\keyword{the}
\keyword{to}
\keyword{transition}
