% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_bins.R
\name{to7daybins}
\alias{to7daybins}
\title{Converts a list time bins rules to a table of rules for each day of the week}
\usage{
to7daybins(rules)
}
\arguments{
\item{rules}{...}
}
\value{
...
}
\description{
Converts a list time bins rules to a table of rules for each day of the week
}
\details{
...
}
\keyword{7}
\keyword{\code{to7daybins}}
\keyword{a}
\keyword{created}
\keyword{day.}
\keyword{internal}
\keyword{of}
\keyword{per}
\keyword{rows,}
\keyword{rules}
\keyword{table}
\keyword{with}
