% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traveltimeHMM.R
\name{traveltimeHMM}
\alias{traveltimeHMM}
\title{\code{traveltimeHMM} estimates trips and link specific speed parameters from observed average speeds per unique trip and link.}
\usage{
traveltimeHMM(logspeeds, trips, timeBins, linkIds, nQ = 1L,
  model = c("HMM", "trip-HMM", "trip", "no-dependence"), tol.err = 10,
  L = 10L, max.it = 20, verbose = FALSE, maxSpeed = NULL,
  seed = NULL, transition_matrix = NULL, initial_state_values = NULL)
}
\arguments{
\item{logspeeds}{A numeric vector of speed observations (in km/h) on the (natural) log-scale.}

\item{trips}{An integer or character vector of trip ids for each observation of \code{speed}.}

\item{timeBins}{A character vector of time bins for each observation of \code{speed}.}

\item{linkIds}{A vector of link ids (route or way) for each observation of \code{speed}.}

\item{nQ}{Integer number of states, default is \code{1}.}

\item{model}{Type of model as string, \code{trip-HMM} to use a hidden Markov model (HMM) with trip effect, \code{HMM} (default) is an HMM without trip effect,
\code{trip} is trip effect model without HMM, and \code{no-dependence} is model with link specific parameter only without an HMM nor a trip effect.}

\item{tol.err}{A numeric variable representing the level of tolerable distance between parameter estimates from consecutive iterations.}

\item{L}{An integer minimum number of observations per factor (\code{linkIds x timeBins}) to estimate the parameter for, \code{default = 10}. Factors that have less observations than \code{L} their estimates are imputed by the average over timeBins.}

\item{max.it}{An integer for the maximum number of iterations to run for, default = \code{NULL}.}

\item{maxSpeed}{A float for the maximum speed in km/h, on the linear scale (not the log-scale, unlike for \code{logspeeds}).}

\item{...}{extra specific parameters, see details}
}
\value{
\code{traveltimeHMM} returns a list of the following parameters

\item{factors}{a factor of interactions (linkId x timeBin) with the same length of observations, and with levels corresponding to unique factors}
\item{trip}{a factor of trips.}
\item{tmat}{a transition matrix with rows corresponding to \code{levels(factors)}, with columns being the row-wise transition matrix of that factor. For example, \code{matrix(tmat[1,], ncol = nQ, nrow = nQ, byrow = TRUE)} is the transition matrix of \code{levels(factors)[1]}.}
\item{init}{a initial state probability matrix with rows corresponding to \code{levels(factors)}, and columns to the \code{nQ} states.}
\item{sd}{a matrix of standard deviations estimates with rows corresponding to \code{levels(factors)}, and columns to standard deviation estimates of the \code{nQ} states.}
\item{mean}{a matrix of mean estimates with rows corresponding to \code{levels(factors)}, and columns to mean estimates for the \code{nQ} states.}
\item{tau}{a numeric variable for the standard deviation estimate of the trip effect parameter \code{E}.}
\item{E}{a numeric vector of trip effect estimates corresponding to \code{levels(trip)}.}
\item{nQ}{an integer number of states.}
\item{nB}{an integer number of unique time bins.}
\item{nObs}{an integer number of observations.}
\item{model}{the type of model used.}
}
\description{
\code{traveltimeHMM} estimates trips and link specific speed parameters from observed average speeds per unique trip and link.
}
\examples{
\dontrun{
data(tripset)
?traveltimeHMM  # for help
fit <- traveltimeHMM(tripset$logspeed, tripset$trip, tripset$timeBin, tripset$linkId, nQ = 2, max.it = 2)
single_trip <- subset(tripset, trip==2700)
pred <- predict.traveltime(fit, single_trip,single_trip$time[1])
hist(pred)      # histogram of prediction samples
mean(pred)      # travel time point estimate
sum(single_trip$traveltime)    # observed traveltime
}
}
