% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{cleanGPS_trip}
\alias{cleanGPS_trip}
\title{\code{cleanGPS} cleans raw GPS data}
\usage{
cleanGPS_trip(trip, trip_id = NULL, verbose = FALSE, ...)
}
\description{
\code{cleanGPS} cleans raw GPS data
}
\details{
Cleans raw GPS data by removing trips with outside the indicated parameters.
}
