% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forwardback.R
\name{forwardback}
\alias{forwardback}
\title{Computing the posterior marginals of hidden state variables using a sequence of observations}
\usage{
forwardback(probTran, initQ)
}
\arguments{
\item{probTran}{An \code{nObs x nQ^2} matrix having each row represent 
\eqn{P(Obs_k | state_k ) * P(state_k | state_{k-1})}.  Each row of \code{probTran} is such that
\code{matrix(probTran[1,], ncol=nQ, nrow=nQ, byrow =TRUE)} is the probability above in matrix format.}

\item{initQ}{A \code{nQ}-sized vector that represents the initial probability of the first link in a trip.}
}
\value{
\code{probTran} returns a list of \code{alpha} and \code{beta}
(respectively forward and backward probability row-wise vectors) such
that \code{normalize(alpha[i] * beta[i])} is the smoothed HMM state estimate.
}
\description{
Computing the posterior marginals of hidden state variables using a sequence of observations
}
\references{
{Russell, S., Norvig, P., 2009.  Artificial Intelligence: A Modern Approach.  Pearson Education, UK.}
}
\keyword{1}
\keyword{3}
\keyword{Algorithm}
\keyword{Norvig's}
\keyword{Russell}
\keyword{Woodard}
\keyword{\code{forwardback}}
\keyword{al.}
\keyword{algorithm}
\keyword{and}
\keyword{at}
\keyword{et}
\keyword{forward-backward}
\keyword{implements}
\keyword{in}
\keyword{internal}
\keyword{is}
\keyword{of}
\keyword{step}
\keyword{that}
\keyword{used}
