% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm_estimates.R
\name{initial_est}
\alias{initial_est}
\title{Commpute initial state probabilites}
\usage{
initial_est(state_prob, by_factor, subset = NULL)
}
\arguments{
\item{state_prob}{An \code{n x Q} matrix of initial state probabilities for \code{n} observations and \code{Q} states}

\item{by_factor}{A vector of factors (link + time bin) of size \code{n} (a factor for each row of \code{state_prob}).}

\item{subset}{A vector indexing the rows of \code{state_prob} for which states probabilities are computed using only this subset, defualt is \code{NULL}}
}
\value{
An \code{m x Q} matrix of probabilites for the m levels of \code{by_factor}.
}
\description{
Commpute initial state probabilites
}
\examples{
\dontrun{
x = runif(10)
x = cbind(x, 1-x)
initial_est(x, factor(sample(c("A", "B"), 10, replace=TRUE)))
}
}
\references{
{Woodard, D., Nogin, G., Koch, P., Racz, D., Goldszmidt, M., Horvitz, E., 2017.  Predicting travel time reliability using mobile phone GPS data.  Transportation Research Part C, 75, 30-44.}
}
\keyword{1}
\keyword{2017}
\keyword{4}
\keyword{Algorithm}
\keyword{Small_Gamma_j,b^(t+1)(q)}
\keyword{These}
\keyword{Woodard}
\keyword{\code{by_factor},}
\keyword{\code{initial_est}}
\keyword{a.,}
\keyword{as}
\keyword{at}
\keyword{correspond}
\keyword{defined}
\keyword{each}
\keyword{et}
\keyword{expectation}
\keyword{for}
\keyword{in}
\keyword{initial}
\keyword{internal}
\keyword{level}
\keyword{level.}
\keyword{of}
\keyword{passed}
\keyword{probabilities}
\keyword{return}
\keyword{state}
\keyword{step}
\keyword{the}
\keyword{to}
\keyword{vectors}
