% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm_estimates.R
\name{initial_est}
\alias{initial_est}
\title{Commputes initial state probabilites 
\code{initial_est} return the state probabilities for each level in the passed \by{by_factor}.}
\usage{
initial_est(state_prob, by_factor, subset = NULL)
}
\arguments{
\item{state_prob}{An \code{n x Q} martrix of initial state probabilities for \code{n} observations and \code{Q} states}

\item{by_factor}{A vector of factors of size \code{n} (a factor for each row of \code{state_prob}).}

\item{subset}{A vector indexing the rows of \code{state_prob} for which states probabilities are computed using only this subset, defualt is \code{NULL}}
}
\value{
An \code{m x Q} matrix of probabilites for the m levels of \code{by_factor}.
}
\description{
Commputes initial state probabilites 
\code{initial_est} return the state probabilities for each level in the passed \by{by_factor}.
}
\examples{
\dontrun{
x = runif(10)
x = cbind(x, 1-x)
initial_est(x, factor(sample(c("A", "B"), 10, replace=TRUE)))
}
}
