% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.traveltime}
\alias{predict.traveltime}
\title{Predict the travel time for a trip using a \code{traveltimeHMM} model object}
\usage{
\method{predict}{traveltime}(modelObject, tripdata,
  starttime = Sys.time(), nbRuns = 1000, logE = NULL)
}
\arguments{
\item{modelObject}{A list provided through the execution of function \code{timetravelHMM}.
The list includes information on model as well as estimates for its parameters.
See \code{timetravelHMM} man page.}

\item{tripdata}{A data frame of road links with information on each
link's traversal.  Columns minimally includes objects 'linkID' and 'length',
and the latter must have the same length.  Rows must be in chronological order.}

\item{starttime}{The start date and time for the very first link of the trip,
in POSIXct format.  Default is the current date and time.}

\item{nbRuns}{Number of simulation runs.  Default is 1000.}

\item{logE}{Vector of trip effects.  Incomplete, to be checked.  ÉG 2019/07/24}
}
\value{
\code{predict.traveltime} returns a vector of size nbRuns of numerics representing the point prediction of total travel time, in seconds, for each run.
}
\description{
\code{predict.traveltime} performs a point prediction by simulation using parameter estimates provided by a \code{traveltimeHMM} model object.
}
\examples{
\dontrun{
data(tripset)
?traveltimeHMM  # for help
fit <- traveltimeHMM(tripset$logspeed, tripset$tripID, tripset$timeBin, tripset$linkID, nQ = 2, max.it = 2)
single_trip <- subset(tripset, tripID==2700)
pred <- predict.traveltime(fit, single_trip,single_trip$time[1])
hist(pred)      # histogram of prediction samples
mean(pred)      # travel time point estimate
sum(single_trip$traveltime)    # observed traveltime
}

}
\references{
{Woodard, D., Nogin, G., Koch, P., Racz, D., Goldszmidt, M., Horvitz, E., 2017.  Predicting travel time reliability using mobile phone GPS data.  Transportation Research Part C, 75, 30-44.}
}
