% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traveltimeHMM.R
\name{traveltimeHMM}
\alias{traveltimeHMM}
\title{Estimate trip- and link- specific speed parameters from observed average speeds}
\usage{
traveltimeHMM(logspeeds = NULL, trips = NULL, timeBins = NULL,
  linkIds = NULL, data = NULL, nQ = 1L, model = c("HMM",
  "trip-HMM", "trip", "no-dependence"), tol.err = 10, L = 10L,
  max.it = 20L, verbose = FALSE, max.speed = NULL, seed = NULL,
  tmat.p = NULL, init.p = NULL)
}
\arguments{
\item{logspeeds}{A numeric vector of speed observations (in km/h) on the (natural) log-scale.  Needs
to be provided if \code{data} is \code{NULL}, otherwise it will be overwritten by \code{logspeeds} column in \code{data}.  Default is \code{NULL}.}

\item{trips}{An integer or character vector of trip ids for each observation of \code{speed}.  Needs
to be provided if \code{data} is \code{NULL}, otherwise it will be overwritten by \code{trips} column in \code{data}. Default is \code{NULL}.}

\item{timeBins}{A character vector of time bins for each observation of \code{speed}. Needs
to be provided if \code{data} is \code{NULL}, otherwise it will be overwritten by \code{timeBins} column in \code{data}. Default is \code{NULL}.}

\item{linkIds}{A vector of link ids (route or way) for each observation of \code{speed}.  Needs
to be provided if \code{data} is \code{NULL}, otherwise it will be overwritten by \code{linkIds} column in \code{data}. Default is \code{NULL}.}

\item{data}{A data frame or equivalent object that contains one column for each of
\code{logspeeds}, \code{trips}, \code{timeBins} and \code{linkIds}.  Default is \code{NULL}.  This
argument is mutually exclusive with the full joint set of \code{logspeeds}, \code{trips},
\code{timeBins} and \code{linkIds}. One can either pass the latter vectors explicitly or as a dataframe \code{data}.}

\item{nQ}{An integer corresponding to the number of different congestion states that the traversal
of a given link can take corresponding to \code{|{1...., Q}|}.  Hidden Markov models (see table 2)
require \code{nQ >= 2} whilst other models require exactly \code{nQ = 1}.  As an example,
Woodard uses \code{nQ = 1}.  Default is \code{1}.}

\item{model}{Type of model as string, \code{trip-HMM} to use a hidden Markov model (HMM) with trip effect, \code{HMM} (default) is an HMM without trip effect,
\code{trip} is trip effect model without HMM, and \code{no-dependence} is model with link specific
parameter only without an HMM nor a trip effect.}

\item{tol.err}{A numeric variable representing the threshold under which the estimation algorithm will
consider it has reached acceptable estimate values.  Default is \code{10}.}

\item{L}{An integer minimum number of observations per factor (\code{linkIds x timeBins}) to estimate
the parameter for.  Default is \code{10}. Factors that have fewer total observations or initial state
observations than \code{L} have their estimates imputed using time bin data, without regard to road
link data.}

\item{max.it}{An integer for the upper limit of the iterations to be performed by the estimation
algorithm.  Default is \code{20}.}

\item{verbose}{A boolean that triggers verbose output.  Default is \code{FALSE}.}

\item{max.speed}{An optional float for the maximum speed in km/h, on the linear scale
(not the log-scale, unlike for \code{logspeeds}).  Default is \code{NULL} which results in a maximum speed
of 130 km/h.  If there is any occurrence of speed in excel of \code{max.speed}, then the system issues to the
user a warning that includes the percentage of such occurrences.}

\item{seed}{An optional float for the seed used for the random generation of the first Markov transition matrix
and initial state vector.  Default is \code{NULL}.  If not provided, then those objects are generated deterministically.
The effect of \code{seed} is cancelled by tmat.p or init.p when provided.}

\item{tmat.p}{An optional starting value for the Markov transition matrix \eqn{\Gamma} of size \code{nQ x nQ}
with rows summing to \code{1}.  Default is \code{NULL}}

\item{init.p}{An optional starting value for the Markov initial state vector \eqn{\gamma}
of size \code{nQ} with elements summing to \code{1}.  Default is \code{NULL}.}
}
\value{
\code{traveltimeHMM} returns a list of the following parameters.
\item{factors}{a factor of interactions (linkId x timeBin) of length \code{nObs}.  Factors are in the format \code{linkId.timeBin}.}
\item{trip}{a factor of trip IDs.}
\item{tmat}{a transition matrix with rows corresponding to \code{levels(factors)}, and with columns being the row-wise transition matrix of that factor. For example, \code{matrix(tmat[1,], ncol = nQ, nrow = nQ, byrow = TRUE)} is the transition matrix of \code{levels(factors)[1]}.  \code{NULL} if hidden Markov modelling is not handled by the selected model type.}
\item{init}{a initial state probability matrix with rows corresponding to \code{levels(factors)}, and columns to the \code{nQ} states.  \code{NULL} if hidden Markov modelling is not handled by the selected model type.}
\item{sd}{a matrix of standard deviations estimates for the natural logarithm of the speed (in km/h), with rows corresponding to \code{levels(factors)}, and columns to standard deviation estimates for the \code{nQ} states.}
\item{mean}{a matrix of mean estimates for the natural logarithm of the speed (in km/h), with rows corresponding to \code{levels(factors)}, and columns to mean estimates for the \code{nQ} states.}
\item{tau}{a numeric variable for the standard deviation estimate for the trip effect parameter \code{logE}.   Equals \code{1} if trip effect is not handled by the selected model type.}
\item{logE}{a numeric vector of trip effect estimates corresponding to \code{levels(trip)}.  Values are set to \code{0} if trip effect is not handled by the selected model type. Units are the same as for \code{logspeeds}.}
\item{nQ}{an integer corresponding to the number of different congestion states, equal to the parameter nQ that was passed in the function call.}
\item{nB}{an integer number of unique time bins.}
\item{nObs}{an integer number of observations.}
\item{model}{a character string corresponding to the type of model used.}
}
\description{
\code{traveltimeHMM} estimates trip- and link- specific speed parameters from observed average speeds per unique trip and link.
}
\examples{
\dontrun{
data(tripset)

# Fit an HMM model with 2 hidden congestion states and 20 algorithm iterations
fit <- traveltimeHMM(tripset$logspeed, tripset$tripID, tripset$timeBin, 
                      tripset$linkID, nQ = 2, max.it = 20)

# Perform prediction - use ?predict.traveltime for details
single_trip <- subset(tripset, tripID==2700)
pred <- predict.traveltime(fit, single_trip,single_trip$time[1])
hist(pred)
mean(pred)
sum(single_trip$traveltime)

?traveltimeHMM      # for help on traveltimeHMM, the estimation function
?predict.traveltime # for help on predict.traveltime, the prediction function
}

}
\references{
{Woodard, D., Nogin, G., Koch, P., Racz, D., Goldszmidt, M., Horvitz, E., 2017.  Predicting travel time reliability using mobile phone GPS data.  Transportation Research Part C, 75, 30-44.}
}
