% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traveltimeHMM.R
\name{traveltimeHMM}
\alias{traveltimeHMM}
\title{Commputes initial state probabilites 
\code{initial_est} return the state probabilities for each level in the passed \by{by_factor}.}
\usage{
traveltimeHMM(logspeeds, trips, timeBins, linkIds, nQ = 1L,
  model = c("HMM", "trip-HMM", "no-dependence"), tolErr = 1, L = 10,
  max_iter = NULL, verbose = TRUE)
}
\arguments{
\item{logspeeds}{}

\item{nQ}{Integer number of states}

\item{model}{}

\item{verbose}{}
}
\value{

}
\description{
Commputes initial state probabilites 
\code{initial_est} return the state probabilities for each level in the passed \by{by_factor}.
}
\examples{
\dontrun{
}
}
