% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traveltimeHMM.R
\name{traveltimeHMM}
\alias{traveltimeHMM}
\title{\code{traveltimeHMM} estimates trips and link specific speed parameters from observed average speeds per unique trip and link.}
\usage{
traveltimeHMM(speeds, trips, timeBins, linkIds, nQ = 1L,
  model = c("HMM", "trip-HMM", "trip", "no-dependence"), tol.err = 10,
  L = 10L, max.it = NULL, ...)
}
\arguments{
\item{speeds}{A numeric vector of speed observations on the log-scale.}

\item{trips}{An integer or character vector of trip ids for each observation of \code{speed}.}

\item{timeBins}{A character vector of time bins for each observation of \code{speed}.}

\item{linkIds}{A vector of link ids (route or way) for each observation of \code{speed}.}

\item{nQ}{Integer number of states, default is \code{1}.}

\item{model}{Type of model as string, \code{trip-HMM} to use a hidden Markov model (HMM) with trip effect, \code{HMM} (default) is an HMM without trip effect,
\code{trip} is trip effect model without HMM, and \code{no-dependence} is model with link specific parameter only without an HMM nor a trip effect.}

\item{tol.err}{A numeric variable representing the level of tolerable distance between parameter estimates from consecutive iterations.}

\item{L}{An integer minimum number of observations per factor (\code{linkIds x timeBins}) to estimate the paramter for, \code{default = 10}. Factors that have less observations than \code{L} their estimates are imputed by the average over timeBins.}

\item{max.it}{An Integer for the maximum number of iterations to run for, default = \code{NULL}.}

\item{...}{extra specific parameters, see details}
}
\description{
\code{traveltimeHMM} estimates trips and link specific speed parameters from observed average speeds per unique trip and link.
}
\details{

}
\examples{
\dontrun{
}
}
